requirements.txt
aws-cdk-lib==2.122.0
constructs>=10.0.0,<11.0.0
click
attrs==23.1.0
boto3==1.28.3
botocore==1.31.3
certifi==2023.5.7
charset-normalizer==3.2.0
exceptiongroup==1.1.2
h11==0.14.0
idna==3.4
jmespath==1.0.1
outcome==1.2.0
packaging==23.1
prompt-toolkit
Pygments==2.15.1
PyInquirer==1.0.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-lambda-local==0.1.13
regex==2023.6.3
requests==2.31.0
s3transfer==0.6.1
selenium==3.141.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.65.0
trio==0.22.2
trio-websocket==0.10.3
urllib3==1.26.16
wcwidth==0.2.6
webdriver-manager==3.8.6
wsproto==1.2.0
